AI_Chatbot/requirements.txt
flask
python-dotenv
google-generativeai
duckduckgo-search
